refactored descriptors
